Generalize solver architecture, move solver handling to BaseDiscretization, move BaseDiscretization and BaseSource out of meta.
